Moved arg_split to _process_win32.py and _process_posix.py.
